Simplified `nextMove` to avoid nested conditionals
This is a good general pattern for acheiving the same logical flow as several nested conditional statements.  Here we simply return out of the function early if the game is over, otherwise we return another ternary operation to fire the correct play function.

Did I mention that I _love_ ternary returns?