Fixes win / draw announce bug.
In `checkWinner` the code to update `p#player` with the result message
was broken.  Basically it's because after the last `checkWinner` runs,
ending the game there is still a call to `switchPlayer`.  We simply
added the same check to `gameOver` as is exiting from the `nextMove`
function.  This restores the expected behavior.